refactor: 출력 뷰가 방대해짐에 따라 결과 뷰로 일부 분리
